requirements.txt
-i https://pypi.org/simple
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
google-api-core==2.19.2; python_version >= '3.7'
google-auth==2.34.0; python_version >= '3.7'
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-storage==2.18.2; python_version >= '3.7'
google-crc32c==1.6.0; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos==1.65.0; python_version >= '3.7'
idna==3.8; python_version >= '3.6'
proto-plus==1.24.0; python_version >= '3.7'
protobuf==5.28.0; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
sh==2.0.7; python_version < '4.0' and python_full_version >= '3.8.1'
urllib3==2.2.2; python_version >= '3.8'
